pyproject.toml
[project]
name = "scythe-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "scythe-engine>=0.0.1",
]
